Major cleanup of main() function
* Much of the main method broken up in to smaller functions to make it
significantly more readable & unit testable.
* Added docstrings to remaining undocumented functions and improved
existing ones.
* Added a try except to creating the configuration directory so that the
script does not crash if download album art cannot be cached.
* Created a Song named tuple to store name information.

+++ spotify.r.1s+.py
@@ -1,51 +1,78 @@
 #!/usr/bin/env python3
-# Spotify argos extension
-# by Jamie Luckett https://github.com/jamieluckett
+"""
+Spotify media controller argos extension for GNOME 3.
+Documentation and source-code (you're already looking at the latter) can be found at:
+https://github.com/jamieluckett/ArgosSpotifyMpris
+"""
 
 import base64
 import hashlib
 import os
 import traceback
+from collections import namedtuple
 
+import gi
 import requests
-from gi.repository.GLib import GError
 from pydbus import SessionBus
 
-BUS_NAME = "org.mpris.MediaPlayer2.spotify"
-OBJECT_PATH = "/org/mpris/MediaPlayer2"
+# USER CONFIGURATION ############################################################################
+# Directory to cache previously loaded album art in                                             #
+IMAGE_CACHE_DIR = ".config/argos/spotify/"                                                      #
+# Icon definitions - icons can be the icon of a program or any built-in GNOME icon              #
+SPOTIFY_ICON = "spotify-client"                                                                 #
+PAUSE_ICON = "media-playback-pause"                                                             #
+PLAY_ICON = "media-playback-start"                                                              #
+STOP_ICON = "media-playback-stop"                                                               #
+BACKWARD_ICON = "media-skip-backward"                                                           #
+FORWARD_ICON = "media-skip-forward"                                                             #
+# Colour of the unclickable song information lines                                              #
+UNCLICKABLE_COLOUR = "#888888"                                                                  #
+# Album Art width (albums are usually square so this is also the height!)                       #
+IMAGE_WIDTH = 250                                                                               #
+# END OF USER CONFIGURATION #####################################################################
 
-IMAGE_CACHE_DIR = ".config/argos/spotify/"
-
-SPOTIFY_ICON = "spotify-client"
-PAUSE_ICON = "media-playback-pause"
-PLAY_ICON = "media-playback-start"
-STOP_ICON = "media-playback-stop"
-BACKWARD_ICON = "media-skip-backward"
-FORWARD_ICON = "media-skip-forward"
-
+# Dictionary mapping current song state to the icon appearing in the ticker
 PLAYBACK_STATUS_ICON_MAPPING = {
     "Paused": PAUSE_ICON,
     "Playing": PLAY_ICON,
     "Stopped": STOP_ICON
 }
 
-UNCLICKABLE_COLOUR = "#888888"
+BUS_NAME = "org.mpris.MediaPlayer2.spotify"  # Spotify dbus bus name
+OBJECT_PATH = "/org/mpris/MediaPlayer2"
+
+Song = namedtuple("Song", "title primary_artist playback_status art_url artist_list album_name")
+
+
+class NoSongException(Exception):
+    pass
 
 
 def debug_print(body):
-    """Prints debug messages if the environment flag DEBUG is 1."""
+    """Prints debug messages if the environment flag "DEBUG" is 1."""
     if os.environ.get("DEBUG") == "1":
         print(body)
 
 
-def make_image_cache_dir():
-    """Creates IMAGE_CACHE_DIR if it doesn't exist yet."""
-    os.makedirs(IMAGE_CACHE_DIR, exist_ok=True)
+def make_image_cache_dir(directory_path=IMAGE_CACHE_DIR):
+    """Creates the image cache directory (IMAGE_CACHE_DIR) if it doesn't exist yet.
+    :return: bool: Whether the image cache directory exists or not.
+    """
+    try:
+        os.makedirs(directory_path, exist_ok=True)
+        return True
+    except Exception:
+        debug_print("Failed to create Image Cache directory {0}".format(directory_path))
+        # For some reason the above makedirs failed (even with exist_ok set to True)
+        # So we'll just check that directory_path is a directory and continue...
+        return os.path.isdir(directory_path)
 
 
 def argos_print(body="", **kwargs):
     """
-    :type body: str
+    Generic function to print nicely in the correct argos format the passed body and arguments.
+    :param body: str: The text to print on the line
+    :param kwargs: {String, String} mappings to be appended to the body as Argos properties
     """
     def arg_format(arg):
         """Reads in passed argument and generates line to put after pipe in argos.
@@ -53,6 +80,7 @@ def argos_print(body="", **kwargs):
         name = arg[0]
         value = arg[1]
         if type(value) == str and ' ' in value:
+            # Arguments with spaces in must be wrapped in quotation marks
             return "{0}='{1}'".format(name, value)
         return "{0}={1}".format(name, value)
 
@@ -64,7 +92,7 @@ def argos_print(body="", **kwargs):
 
 
 def print_argos_separator():
-    """Prints an argos line separator (---)"""
+    """Prints an argos line separator."""
     print("---")
 
 
@@ -78,8 +106,9 @@ def get_spotify_object():
 def get_art(art_url):
     """Returns a base64 encoded jpeg of the album art at art_url.
     Saves all downloaded images to IMAGE_CACHE_DIR under the name of art_url md5 hashed.
-    :type art_url: string
-    :rtype string
+    Checks the IMAGE_CACHE_DIR first, prioritising any image found there over downloading a new one.
+    :param art_url: str: URL of album art to get.
+    :return str: Returns the album art after being base64 encoded.
     """
     art_url_hash = hashlib.md5(art_url.encode('utf-8')).hexdigest()
     image_location = IMAGE_CACHE_DIR + art_url_hash
@@ -93,15 +122,16 @@ def get_art(art_url):
         b64_encoded = base64.b64encode(image)
         # TODO - Find a nicer way to return this byte object as a string without b'
         b64_string = repr(b64_encoded)[1:]
-        make_image_cache_dir()
-        with open(image_location, 'w') as f:
-            f.write(b64_string)
+        dir_exists = make_image_cache_dir(IMAGE_CACHE_DIR)
+        if dir_exists:
+            with open(image_location, 'w') as f:
+                f.write(b64_string)
         return b64_string
 
 
 def print_control_menu(playback_status):
     """Prints the music control menu options.
-    :type playback_status: str
+    :param playback_status: str: Status of the current song's playback.
     """
     if playback_status in ["Stopped", "Paused"]:
         argos_print("<b>Play</b>/Pause",
@@ -128,47 +158,85 @@ def print_control_menu(playback_status):
                 terminal="false")
 
 
-if __name__ == "__main__":
-    try:
-        spotify_object = get_spotify_object()
-        metadata = spotify_object.Metadata
-        debug_print(metadata)
-        artist_list = metadata['xesam:artist']
-        if artist_list:
-            # Get current song information
-            primary_artist = metadata['xesam:artist'][0]
-            playback_status = spotify_object.PlaybackStatus
-            song_title = metadata['xesam:title']
-            art_url = metadata['mpris:artUrl']
-            album_name = metadata['xesam:album']
-            # Print current song
-            argos_print("{0} - {1}".format(primary_artist, song_title),
-                        iconName=PLAYBACK_STATUS_ICON_MAPPING[playback_status],
-                        useMarkup="false",
-                        unescape="true")
-            print_argos_separator()
-            argos_print("Song Title: {0}".format(song_title), color=UNCLICKABLE_COLOUR)
-            argos_print("Album: {0}".format(album_name), color=UNCLICKABLE_COLOUR)
-            if song_title:
-                # If only the plural of Artist was Artist...
-                # Spotify never seems to actually return more than 1 artist through mpris though but maybe one day
-                argos_print("Artists: {0}".format(", ".join(artist_list)), color=UNCLICKABLE_COLOUR)
-            else:
-                argos_print("Artist: {0}".format(artist_list[0]), color=UNCLICKABLE_COLOUR)
-            print_control_menu(playback_status)
-            argos_print(image=get_art(art_url), imageWidth="200")
-        else:
-            argos_print("Nothing Playing", iconName=STOP_ICON)
+def get_current_song():
+    """
+    Gets information on the current playing song from the DBus, raises a SpotifyClosedException if it cannot.
+    :return Song: A Song object created with song attributes pulled from the dbus.
+    """
+    # Get Spotify data from dbus
+    spotify_object = get_spotify_object()
+    metadata = spotify_object.Metadata
+    debug_print(metadata)
 
-    except GError as e:
+    # Try and return a Song object, otherwise give up and admit that no song is playing
+    try:
+        return Song(
+            metadata['xesam:title'],
+            metadata['xesam:artist'][0],
+            spotify_object.PlaybackStatus,
+            metadata['mpris:artUrl'],
+            metadata['xesam:artist'],
+            metadata['xesam:album']
+        )
+    except KeyError:
+        raise NoSongException
+
+
+def print_song(song):
+    """
+    Prints the current song's information and album art, along with the relevant controls for the playback state.
+    :param song: Song: Song to print to the navbar.
+    """
+    argos_print("{0} - {1}".format(song.primary_artist, song.title),
+                iconName=PLAYBACK_STATUS_ICON_MAPPING[song.playback_status],
+                useMarkup="false",
+                unescape="true")
+    print_argos_separator()
+    argos_print("Song Title: {0}".format(song.title), color=UNCLICKABLE_COLOUR, useMarkup="false")
+    argos_print("Album: {0}".format(song.album_name), color=UNCLICKABLE_COLOUR, useMarkup="false")
+
+    if len(song.artist_list) > 1:
+        # Spotify never seems to actually return more than 1 artist through mpris, but maybe one day...
+        argos_print("Artists: {0}".format(", ".join(song.artist_list)),
+                    color=UNCLICKABLE_COLOUR, useMarkup="false")
+    else:
+        argos_print("Artist: {0}".format(song.artist_list[0]), color=UNCLICKABLE_COLOUR, useMarkup="false")
+    print_control_menu(song.playback_status)
+    # TODO - Make clicking the album art bring spotify to the foreground.
+    argos_print(image=get_art(song.art_url), imageWidth=IMAGE_WIDTH)
+
+
+def print_last_exception():
+    """
+    Prints the last thrown Exception out.
+    Also debug_prints it in case the issue lives in argos_print.
+    """
+    last_exception = traceback.format_exc()
+    debug_print(last_exception)
+    argos_print("Something went wrong!", iconName="dialog-warning")
+    print_argos_separator()
+    argos_print("<b>Exception Details:</b>")
+    argos_print(last_exception.replace("\n", "\\n"), font="monospace", useMarkup="false", unescape="true")
+
+
+def main():
+    """
+    Main function for SpotifyArgos.
+    """
+    try:
+        current_song = get_current_song()
+        print_song(current_song)
+    except NoSongException:
+        argos_print("Nothing Playing", iconName=STOP_ICON)
+    except gi.repository.GLib.GError:
         # A GError indicates that (most likely) Spotify isn't open.
         argos_print("Spotify isn't open!", iconName=STOP_ICON)
         print_argos_separator()
         argos_print("Open Spotify", iconName=SPOTIFY_ICON, bash="spotify U%", terminal="false")
-    except Exception as e:
+    except Exception:
         # Catch generic exception here so it can be displayed in the dropdown
-        argos_print("Something went wrong!", iconName="dialog-warning")
-        print_argos_separator()
-        argos_print("<b>Exception Details:</b>")
-        tb = traceback.format_exc()
-        argos_print(tb.replace("\n", "\\n"), font="monospace", useMarkup="false", unescape="true")
+        print_last_exception()
+
+
+if __name__ == "__main__":
+    main()
